Add files via upload
DO NOT COMMIT OR MERGE. 

I have tried to make the image bigger but now the text doesn't look so good. Anyway with this patch you get an idea of how you can try to resize things. I encourage you to try to play with this and you might find a better combination to make the images and text look better.