Add files via upload
Sorry it took me a little while to put the visual example together, this is what I meant.

Take a look at the Top 5 subsections, each activity has a carousel of 3 images. I thought this is what I was agreeing to in our meeting while the hero image stayed as a single image. Do we want to add this feature? I think it makes the page less static. 

-Shara